+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 4.18.5.1
 
 - Hello, Color Dialog! Sample
+- Hello, Font Dialog! Sample
 - Handle SWT RGB color data objects when setting colors on widgets (e.g. background)
 - Fixed issue with tree multi selection throwing an exception